# src/config/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ExpressJs boilerplate",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://localhost:3000/api"
    }
  ],
  "paths": {
    "/auth": {
      "post": {
        "description": "Login the user",
        "requestBody": {
          "$ref": "#/components/requestBodies/AuthRequest"
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/GeneralError"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "description": "Get user data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "default": {
          "$ref": "#/components/responses/GeneralError"
        }
      }
    },
    "/users": {
      "post": {
        "description": "Register a new user",
        "requestBody": {
          "$ref": "#/components/requestBodies/AuthRequest"
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "default": {
            "$ref": "#/components/responses/GeneralError"
          }
        }
      }
    },
    "/swagger": {
      "x-swagger-pipe": "swagger_raw"
    }
  },
  "components": {
    "schemas": {
      "Auth": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "format": "password"
          }
        }
      },
      "Error": {
        "required": [
          "message",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Object error class"
          },
          "message": {
            "type": "string",
            "description": "Human readable error description"
          }
        }
      }
    },
    "responses": {
      "GeneralError": {
        "description": "Common structure for error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "requestBodies": {
      "AuthRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Auth"
            },
            "examples": {
              "auth": {
                "summary": "User Example",
                "value": {
                  "email": "abc@abc.com",
                  "password": "mySecretPassword"
                }
              }
            }
          }
        },
        "required": true
      },
      "UserRegistrationRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Auth"
            },
            "examples": {
              "auth": {
                "summary": "User Example",
                "value": {
                  "email": "abc@abc.com",
                  "password": "mySecretPassword"
                }
              }
            }
          }
        },
        "required": true
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}